requirements.txt
fastapi
uvicorn[standard]
websockets
google-genai
pydantic
pydantic-settings
python-multipart
python-dotenv
aiofiles
